requirements.txt
langgraph>=0.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0
tavily-python>=0.3.0
python-telegram-bot>=20.0
psutil>=5.9.0
rich>=13.0.0
black>=24.0.0
